Allow the window to become smaller.
The smallest allowed size is actually determined by the smallest allowed size of the drawer. Our drawer automatically closes if the window gets too small for it, so we don't really care about it's minimum height. (We do care about the drawers width, which should be locked to 250px, so we have to set a min and max contents size.)